# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "This is the API documentation for my project.",
        "contact": {
            "email": "developer@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:6162/"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login and get a token",
                "description": "Login with email and password, then retrieve an authentication token.",
                "operationId": "a3b306d14572d1f4bd6c064b3233e7b8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "JWT Token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/translations": {
            "post": {
                "tags": [
                    "Translations"
                ],
                "summary": "Create a new translation",
                "operationId": "ebe13437cd4bd8c2184e448fcfbb01ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "content",
                                    "tags"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "properties": {
                                            "en": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "tags": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Translation created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/translations/{id}": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Get Translation by ID",
                "description": "Retrieve a translation by its ID.",
                "operationId": "cc5e674b8b2eb8f545c16ed997e853b5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Translation found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Translations"
                ],
                "summary": "Update a translation",
                "operationId": "e314c7afaf04580ad65b747c8a1f7ce6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key",
                                    "content",
                                    "tags"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "properties": {
                                            "en": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "tags": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Translation updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Translation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Translation not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/translation/search": {
            "get": {
                "tags": [
                    "Translations"
                ],
                "summary": "Search for translations",
                "description": "Search for translations based on key, tags, or content",
                "operationId": "searchTranslations",
                "parameters": [
                    {
                        "name": "key",
                        "in": "query",
                        "description": "Search translations by key",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "tags",
                        "in": "query",
                        "description": "Search translations by tags",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Translation"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid search parameters."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Translation": {
                "required": [
                    "key",
                    "content",
                    "tags"
                ],
                "properties": {
                    "key": {
                        "description": "The key of the translation",
                        "type": "string"
                    },
                    "content": {
                        "properties": {
                            "en": {
                                "description": "Translation in English",
                                "type": "string"
                            },
                            "fr": {
                                "description": "Translation in French",
                                "type": "string"
                            },
                            "es": {
                                "description": "Translation in Spanish",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "tags": {
                        "description": "Comma-separated list of tags",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Translations",
            "description": "Translations"
        }
    ]
}